feat: refactor checks view into smaller parts
