+++ Pareto/Checks/System Integrity/SecureTerminal.swift
@@ -24,14 +24,6 @@ class SecureTerminalCheck: ParetoCheck {
         "Terminal is not using secure entry"
     }
 
-    override var hasDebug: Bool {
-        true
-    }
-
-    override func debugInfo() -> String {
-        readDefaultsNative(path: "com.apple.Terminal", key: "SecureKeyboardEntry") ?? "No data"
-    }
-
     override func checkPasses() -> Bool {
         if let enabled = readDefaultsNative(path: "com.apple.Terminal", key: "SecureKeyboardEntry") {
             os_log("SecureKeyboardEntry, status %{enabled}s", log: Log.check, enabled)
